casting unsigned int

+++ bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@ int get_bit(unsigned long int n, unsigned int index)
 	unsigned long int shifted_num;
 	int bit_value;
 	
-	if (index < 0)
+	if ((unsigned int)index < 0)
 		return (-1);
 
 	shifted_num = n >> index;
